--- client-nextjs/src/components/Navbar/MenuItem.jsx
@@ -1,10 +1,0 @@
-
-const MenuItem = ({onClick, label}) => {
-    return ( 
-        <div  onClick={onClick}  className="px-4 py-3 hover:bg-neutral-100 transition font-semibold">
-          {label}
-        </div>
-       );
-    }
-     
-    export default MenuItem;
