--- client-nextjs/src/components/Navbar/UserMenu.jsx
@@ -1,93 +1,0 @@
-import { useCallback, useState } from "react";
-import { AiOutlineMenu } from "react-icons/ai";
-import Avatar from "../Avatar";
-import MenuItem from "./MenuItem";
-import Cart from "./Cart";
-const UserMenu = () => {
-  const [isOpen, setIsOpen] = useState(false);
-  const toggleOpen = useCallback(() => {
-    setIsOpen((value) => !value);
-  }, []);
-
-  return (
-    <div className="relative">
-      <div className="flex flex-row items-center gap-3">
-        <Cart/>
-        <div
-          onClick={()=>{}}
-          className="
-            hidden
-            md:block
-            text-sm
-            font-semibold
-            py-3
-            px-4
-            rounded-full
-            transition
-            cursor-pointer
-            text-white
-          "
-        >
-          EventyPro
-        </div>
-        <div
-        onClick={toggleOpen}
-        className="
-            p-4
-            md:py-1
-            md:px-2
-            py-3
-            px-4
-            border-[1px]
-            border-neutral-200
-            flex
-            flex-row
-            items-center
-            gap-3
-            rounded-full
-            cursor-pointer
-            hover:shadow-md
-            transition
-          "
-        >
-          <AiOutlineMenu/>
-          <div className="hidden md:block">
-            <Avatar/>
-          </div>
-        </div>
-      </div>
-      {isOpen && (
-        <div
-          className="
-            absolute
-            rounded-xl
-            shadow-md
-            w-[40vw]
-            md:w-3/4
-            bg-white
-            overflow-hidden
-            right-0
-            top-12
-            text-sm
-          "
-        >
-          <div className="flex flex-col cursor-pointer">
-              <>
-                <MenuItem
-                  label="Login"
-                  onClick={()=>{}}
-                />
-                <MenuItem
-                  label="Sign up"
-                  onClick={()=>{}}
-                />
-              </>
-
-          </div>
-        </div>
-      )}
-    </div>
-   );
-}
-
-export default UserMenu;
